+++ src/pages/_sqill-for-pros.js
@@ -1,14 +1,22 @@
 import { useState } from 'react';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import Image from 'next/image'
+import Image from 'next/image';
 
 import Layout from '@sqill/components/layout';
 import Form from '@sqill/components/form';
 import Button from '@sqill/components/button';
 import Modal from '@sqill/components/modal';
 
-import { Section, Wrapper, ImageWrapper, H1, H2, VideoWrapper, Div } from '@sqill/components/pages/sqill-for-pros.styles';
+import {
+  Section,
+  Wrapper,
+  ImageWrapper,
+  H1,
+  H2,
+  VideoWrapper,
+  Div,
+} from '@sqill/components/pages/sqill-for-pros.styles';
 
 const SqillForPros = () => {
   const { t } = useTranslation('sqillForPros');
@@ -22,17 +30,22 @@ const SqillForPros = () => {
       <Section>
         <Wrapper>
           <H1>
-            <Image src='/images/logo_gradient.svg' alt='sqill logo' width={80} height={80} />
+            <Image src="/images/logo_gradient.svg" alt="sqill logo" width={80} height={80} />
             {t('header.title')}
           </H1>
           <ImageWrapper>
-            <Image src='/images/sqill4props_cta_image.png' alt='sqill logo' width={1350} height={590} />
+            <Image
+              src="/images/sqill4props_cta_image.png"
+              alt="sqill logo"
+              width={1350}
+              height={590}
+            />
             <VideoWrapper>
-              <video autoPlay playsInline loop muted src='/videos/4pros.mp4' />
+              <video autoPlay playsInline loop muted src="/videos/4pros.mp4" />
             </VideoWrapper>
           </ImageWrapper>
           <H2>{t('header.subtitle')}</H2>
-          <Div dangerouslySetInnerHTML={{__html: t('header.description')}} />
+          <Div dangerouslySetInnerHTML={{ __html: t('header.description') }} />
           <Button onClick={() => setShowForm(true)} label={form.button} />
         </Wrapper>
       </Section>
@@ -42,9 +55,9 @@ const SqillForPros = () => {
           <Form
             fields={form.fields}
             button={form.submit}
-            tableName='4pros'
-            success={`${form.feedback.title} ${form.feedback.text}`}
-            error={form.feedback.error}
+            tableName="4pros"
+            successMessage={`${form.feedback.title} ${form.feedback.text}`}
+            errorMessage={form.feedback.error}
           />
         </Modal>
       )}
@@ -54,7 +67,7 @@ const SqillForPros = () => {
 
 export const getStaticProps = async ({ locale }) => ({
   props: {
-    ...await serverSideTranslations(locale, ['menu', 'footer', 'sqillForPros']),
+    ...(await serverSideTranslations(locale, ['menu', 'footer', 'sqillForPros'])),
   },
 });
