+++ src/pages/contacts.js
@@ -1,7 +1,7 @@
 // import { useRef } from 'react';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import Image from 'next/image'
+import Image from 'next/image';
 
 import Layout from '@sqill/components/layout';
 import Form from '@sqill/components/form';
@@ -18,15 +18,15 @@ const Contacts = () => {
       <Section>
         <Wrapper>
           <H1>
-            <Image src='/images/logo_gradient.svg' alt='sqill logo' width={80} height={80} />
+            <Image src="/images/logo_gradient.svg" alt="sqill logo" width={80} height={80} />
             {t('header.title')}
           </H1>
           <Form
             fields={form.fields}
             button={form.button}
-            tableName='sayHi'
-            success={`${form.feedback.title} ${form.feedback.text}`}
-            error={form.feedback.error}
+            tableName="sayHi"
+            successMessage={`${form.feedback.title} ${form.feedback.text}`}
+            errorMessage={form.feedback.error}
           />
         </Wrapper>
       </Section>
@@ -36,7 +36,7 @@ const Contacts = () => {
 
 export const getStaticProps = async ({ locale }) => ({
   props: {
-    ...await serverSideTranslations(locale, ['menu', 'footer', 'contacts']),
+    ...(await serverSideTranslations(locale, ['menu', 'footer', 'contacts'])),
   },
 });
