+++ src/translations/en/footer.json
@@ -1,33 +1,18 @@
 {
   "description": "This is a non-page component that requires its own namespace",
   "copyright": "sqill (Success Invitation Lda.)",
-  "country": {
-    "title": "Country",
-    "options": [
-      { "label": "Portugal" },
-      { "label": "Brasil" }
-    ]
-  },
   "about": {
     "title": "About",
     "options": [
-      { "label": "About us", "url": "ABOUT" },
       { "label": "Blog", "url": "BLOG" },
-      { "label": "FAQs", "url": "FAQS" },
-      { "label": "Terms and Conditions", "url": "TERMS_AND_CONDITIONS" },
-      { "label": "Privacy Policy", "url": "PRIVACY_POLICY" },
       { "label": "Contacts", "url": "CONTACTS" }
     ]
   },
   "social": {
     "title": "Social",
     "options": [
-      { "label": "Facebook", "url": "FACEBOOK_URL" },
       { "label": "Instagram", "url": "INSTAGRAM_URL" },
-      { "label": "Twitter", "url": "TWITTER_URL" },
-      { "label": "Linkedin", "url": "LINKEDIN_URL" },
-      { "label": "Google Play", "url": "PLAY_STORE_URL" },
-      { "label": "App Store", "url": "APP_STORE_URL" }
+      { "label": "Linkedin", "url": "LINKEDIN_URL" }
     ]
   }
 }
