--- src/translations/en/sqillForBusiness.json
@@ -1,69 +1,0 @@
-{
-  "meta": {
-    "title": "sqill | sqill 4 business",
-    "description": "",
-    "keywords": ""
-  },
-  "header": {
-    "hint": "Try the first month for free",
-    "title": "sqill 4 business",
-    "subtitle": "We enable sports organizations to create the best video content for their social media, helping them connect with a younger born-digital generation.",
-    "description": "Use sqill to optimize your video editing processes and create new content for your social media. The right option for leagues, clubs, and any sports brand.",
-    "hero": {
-      "image": "4biz_hero.png", "width": "804", "height": "464"
-    },
-    "form": {
-      "fields": [
-        { "id": "Name", "label": "Name", "type": "text" },
-        { "id": "About", "label": "Company", "type": "text" },
-        { "id": "Email", "label": "Email", "type": "email" }
-      ],
-      "button": "Let's talk",
-      "submit": "Send",
-      "feedback": {
-        "title": "New registration submitted with success.",
-        "text": "We will send you an email with the next steps",
-        "error": "Something went wrong, please try again"
-      }
-    }
-  },
-  "section1": {
-    "options": [
-      { "key": "1", "title": "Cut long videos in several clips", "description": "Cut your video in seconds, and send it to your mobile with original quality, by using our tool “send2sqill” via a link or upload", "image": "4biz_1.png", "width": "550", "height": "400" },
-      { "key": "2", "title": "Edit your videos - fast and simple", "description": "Transform your horizontal game videos into vertical videos, with an unique dynamic zoom feature. Perfectly suitable for reels and tiktok.", "image": "4biz_2.png", "width": "470", "height": "480" },
-      { "key": "3", "title": "Use our effects and animations with no limitations", "description": "Enhance your video quality with dozens of customized effects. Unlimited use.", "image": "4biz_3.png", "width": "470", "height": "480" },
-      { "key": "4", "title": "Download and share directly into social media", "description": "Save your video on your library to post later or share it directly on your social media.", "image": "4biz_4.png", "width": "470", "height": "480" }
-    ]
-  },
-  "section2": {
-    "title": "Why should you use sqill 4 business?",
-    "options": [
-      { "description": "Enhancement of creativity and personal touch in content creation", "image": "icon_create.png", "width": "100", "height": "104" },
-      { "description": "Very competitive price when compared with alternatives in the market", "image": "icon_euro.png", "width": "100", "height": "100" },
-      { "description": "Simple and intuitive video editor for Mobile", "image": "icon_edit.png", "width": "102", "height": "102" }
-    ],
-    "subsection": {
-      "title": "and…",
-      "options": [
-        { "label": "Immediate signup process" },
-        { "label": "Access to the athletes database" },
-        { "label": "Usage of several “Finder Premium” profiles" }
-      ],
-      "form": {
-        "fields": [
-          { "id": "Name", "label": "Name", "type": "text" },
-          { "id": "About", "label": "Company", "type": "text" },
-          { "id": "Email", "label": "Email", "type": "email" }
-
-        ],
-        "button": "Let's talk",
-        "submit": "Send",
-        "feedback": {
-          "title": "New registration submitted with success.",
-          "text": "We will send you an email with the next steps",
-          "error": "Something went wrong, please try again"
-        }
-      }
-    }
-  }
-}
