+++ src/translations/pt/menu.json
@@ -1,8 +1,7 @@
 {
   "options": [
     { "label": "home", "url": "HOME" },
-    { "label": "atleta", "url": "ATHLETE" },
-    { "label": "sqill for business", "url": "SQILL_FOR_BUSINESS" },
-    { "label": "sqill for pros", "url": "SQILL_FOR_PROS" }
+    { "label": "sqill", "url": "SQILL" },
+    { "label": "sqill para atletas", "url": "SQILL_FOR_ATHLETES" }
   ]
 }
