--- src/translations/pt/sqillForBusiness.json
@@ -1,68 +1,0 @@
-{
-  "meta": {
-    "title": "sqill | sqill 4 business",
-    "description": "",
-    "keywords": ""
-  },
-  "header": {
-    "hint": "EXPERIMENTA GRÁTIS O PRIMEIRO MÊS",
-    "title": "sqill 4 business",
-    "subtitle": "AJUDAMOS AS ORGANIZAÇÕES DESPORTIVAS A CRIAREM MELHOR CONTEÚDO VÍDEO PARA AS SUAS REDES SOCIAIS, CONECTANDO-SE COM UMA NOVA GERAÇÃO BORN-DIGITAL",
-    "description": "USA A SQILL PARA OPTIMIZAR A EDIÇÃO DE VÍDEO E CRIAR NOVOS CONTEÚDOS PARA AS TUAS REDES SOCIAIS. INDICADO PARA LIGAS, CLUBES OU QUALQUER MARCA DESPORTIVA",
-    "hero": {
-      "image": "4biz_hero.png", "width": "804", "height": "464"
-    },
-    "form": {
-      "fields": [
-        { "id": "Name", "label": "Nome", "type": "text" },
-        { "id": "About", "label": "Empresa", "type": "text" },
-        { "id": "Email", "label": "Email", "type": "email" }
-      ],
-      "button": "Vamos falar?",
-      "submit": "Enviar",
-      "feedback": {
-        "title": "Novo registo efectuado com sucesso",
-        "text": "Vamos enviar por email os próximos passos!",
-        "error": "Algo correu mal, por favor tente novamente"
-      }
-    }
-  },
-  "section1": {
-    "options": [
-      { "key": "1", "title": "CORTA VÍDEOS LONGOS EM VÁRIOS CLIPS", "description": "CORTA O VÍDEO EM SEGUNDOS E ENVIA PARA MOBILE COM QUALIDADE ORIGINAL COM A NOSSA FERRAMENTA SEND2SQILL VIA LINK OU UPLOAD", "image": "4biz_1.png", "width": "550", "height": "400" },
-      { "key": "2", "title": "EDITA VÍDEOS COM RAPIDEZ E SIMPLICIDADE", "description": "Transforma vídeos horizontais em vídeos verticais, com a ferramenta única de zoom dinâmico. Perfeito para adaptação ao reels ou tiktok.", "image": "4biz_2.png", "width": "470", "height": "480" },
-      { "key": "3", "title": "USA OS NOSSOS EFEITOS PARA DESPORTOS DE FORMA ILIMITADA", "description": "Melhora os teus vídeos com dezenas de efeitos personalizados. Utiliza de forma ilimitada em quantos vídeos quiseres.", "image": "4biz_3.png", "width": "470", "height": "480" },
-      { "key": "4", "title": "FAZ DOWNLOAD OU PARTILHA DIRETAMENTE NAS REDES SOCIAIS", "description": "Guarda o vídeo para publicar mais tarde ou partilha diretamente nas tuas redes sociais.", "image": "4biz_4.png", "width": "470", "height": "480" }
-    ]
-  },
-  "section2": {
-    "title": "PORQUÊ USAR A SQILL 4 BUSINESS?",
-    "options": [
-      { "description": "POTENCIAÇÃO DE CRIATIVIDADE E PERSONALIDADE NA CRIAÇÃO DE CONTEÚDO", "image": "icon_create.png", "width": "100", "height": "104" },
-      { "description": "PREÇO MUITO ACESSÍVEL FACE A QUALQUER SUBSTITUTO DE MERCADO", "image": "icon_euro.png", "width": "100", "height": "100" },
-      { "description": "EDIÇÃO DE VÍDEO SIMPLES EM MOBILE POR QUALQUER PESSOA", "image": "icon_edit.png", "width": "102", "height": "102" }
-    ],
-    "subsection": {
-      "title": "E ainda…",
-      "options": [
-        { "label": "PROCESSO DE SIGNUP IMEDIATO" },
-        { "label": "ACESSO À BASE DE DADOS DOS ATLETAS" },
-        { "label": "UTILIZAÇÃO DE VÁRIOS PERFIS FINDER PREMIUM" }
-      ],
-      "form": {
-        "fields": [
-          { "id": "Name", "label": "Nome", "type": "text" },
-          { "id": "About", "label": "Empresa", "type": "text" },
-          { "id": "Email", "label": "Email", "type": "email" }
-        ],
-        "button": "Sabe mais",
-        "submit": "Enviar",
-        "feedback": {
-          "title": "Novo registo efectuado com sucesso",
-          "text": "Vamos enviar por email os próximos passos!",
-          "error": "Algo correu mal, por favor tente novamente"
-        }
-      }
-    }
-  }
-}
