+++ src/utils/routes.js
@@ -2,11 +2,12 @@ export default {
   HOME: '/',
   ABOUT: '/about',
   ATHLETE: '/#athlete',
+  SQILL: '/#sqill',
   FAQS: '/faqs',
   CONTACTS: '/contacts',
   BLOG: 'https://app.heysqill.com/blog',
-  SQILL_FOR_BUSINESS: '/sqill-for-business',
   SQILL_FOR_PROS: '/sqill-for-pros',
+  SQILL_FOR_ATHLETES: '/sqill-for-athletes',
   PRIVACY_POLICY: '/privacy-policy',
   TERMS_AND_CONDITIONS: '/terms-and-conditions',
 };
